drivers: ps2: Add resend callback to ps/2 interface
When the PS/2 driver detects an error in an incoming transmission, it’s supposed to send 0xfe to the device to ask it to resend the last packet.

But a PS/2 packet can be more than one byte (such as a 3 or 4-byte mouse packet).

So, on reception of the 0xfe resend command, the PS/2 device resends ALL bytes in the packet and not just the one where the transmission failed.

This can cause the higher level driver’s packet buffer to become misaligned.

This PR adds a callback that notifies the higher level driver when the PS/2 driver requested a resend so that the higher level driver can expect a new packet and doesn’t get misaligned.